cs5542-week5-snowflake---starter/requirements.txt
snowflake-connector-python[secure-local-storage]
pandas
python-dotenv
streamlit
altair
watchdog
faiss-cpu
sentence-transformers
datasets
pytest
pyyaml
